Accept Merge Request #5536: (ly/update-project -> release) Merge Request: 迁移项目设置相关接口 Created By: @刘颖 Reviewed By: @廖凯 Approved By: @廖凯 Accepted By: @刘颖 URL: https://codingcorp.coding.net/p/platform/d/platform-project/git/merge/5536

+++ lib/src/main/java/net/coding/lib/project/entity/ProjectMember.java
@@ -28,7 +28,7 @@ public class ProjectMember implements Serializable {
     /**
      * 类型
      */
-    private Byte type;
+    private Short type;
 
     /**
      * 创建时间
